simplified environment.py Postprocessor by subclassing
